tests/submission/Infraware/Forms/contents/Forms/Not Use/textarea_cols.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>textarea_cols</h3>
	</p>
	
	<li>Cols of textarea control is 5 with respect to the scrollbar.</li>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  id="input_form">
	<p><textarea id='textarea_id'></textarea></p>
	</form>

	<script>
		
		var textarea = document.getElementById("textarea_id");
		textarea.cols = 5;

		if (typeof(textarea.cols) == "number") {
			if (textarea.cols != 5) {
				test( function() {
					assert_true(false, "cols attribute is not correct.");
				});
			}
		} else {
			test( function() {
				assert_unreached("cols attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
